+++ source/lcc/src/dag.c
@@ -54,7 +54,7 @@ static Node node(int op, Node l, Node r, Symbol sym) {
 	int i;
 	struct dag *p;
 
-	i = (opindex(op)^((unsigned long)sym>>2))&(NELEMS(buckets)-1);
+	i = (opindex(op)^((intptr_t)sym>>2))&(NELEMS(buckets)-1);
 	for (p = buckets[i]; p; p = p->hlink)
 		if (p->node.op      == op && p->node.syms[0] == sym
 		&&  p->node.kids[0] == l  && p->node.kids[1] == r)
